requirements.txt
requests>=2.28.0
pdfplumber>=0.10.0
APScheduler>=3.10.0
python-dotenv>=1.0.0
